Move tests to unittests
